feat(cli): modulariza parser, comandos e utilidades
