+++ DUKE/tests/run_tests.cpp
@@ -1,0 +1,11 @@
+#include <cassert>
+
+void test_lerOpcao12();
+void test_adicionarMaterial();
+
+int main() {
+    test_lerOpcao12();
+    test_adicionarMaterial();
+    return 0;
+}
+
